Add files via upload
21 market segments
Comaprison on CoV values
Why APAC_Consumer
Concluding the optimum technique
Comparing the sales forecast plots for all the
smoothing techniques and their MAPE values
Conclusions on which technique works the best for the
sales forecast and why? The reason for this using the
forecast plot and the MAPE values both.